fixed main.cc, equations.cc, symbols.cc and symbols.h
I had to fix issues with the main function of the code, as when you ran it, I had it spam multiple lines and it looked hard to look at. Now it should run properly like a calculator (fixed it so the equations now follow PEMDAS). Need to fix the GUI still as it look horrible and the symbol declaration could use some work.